Move some functions into graph/graph-helpers where they belong
