+++ src/core/data-controller-helpers.js
@@ -15,52 +15,6 @@ function multiYearMeanSelected(props) {
 }
 
 
-function displayError(error, displayMethod) {
-  if (error.response) {
-    // axios error: data server sent a non-200 response
-    displayMethod('Error: ' + error.response.status + ' received from data server.');
-  } else if (error.request) {
-    // axios error: data server didn't respond
-    displayMethod('Error: no response received from data server.');
-  } else {
-    // either an error thrown by a data validation function,
-    // an error thrown by the DataGraph or DataTable parsers,
-    // or the generic and somewhat unhelpful 'Network Error' from axios
-    // Testing turned up 'Network Error' in two cases:
-    // the server is down, or the server has a 500 error.
-    // Other http error statuses tested were reflected in
-    // error.response.status as expected
-    // (see https://github.com/mzabriskie/axios/issues/383)
-    displayMethod(error.message);
-  }
-}
-
-
-function noDataMessageGraphSpec(message) {
-  return {
-    data: {
-      columns: [],
-      empty: {
-        label: {
-          text: message,
-        },
-      },
-    },
-    axis: {},
-  };
-}
-
-
-const blankGraphSpec = {
-  data: {
-    columns: [],
-  },
-  axis: {},
-};
-
 export {
   multiYearMeanSelected,
-  displayError,
-  noDataMessageGraphSpec,
-  blankGraphSpec,
 };
